Moved layer detectors to layers_onnx.h/cpp
