+++ include/eddl/serialization/onnx/export_helpers.h
@@ -10,9 +10,6 @@ onnx::ModelProto build_onnx_model(Net *net, bool gradients);
 // Builds the graph of the ModelProto from the net
 void set_graph(onnx::ModelProto *model, Net *net, bool gradients);
 
-// Builds a node in the onnx graph from the layer of eddl
-void build_node_from_layer(Layer *layer, onnx::GraphProto *graph, bool gradients, bool is_recurrent);
-
 // Fixes the input shape for recurrent models
 void prepare_recurrent_input(string input_name, string output_name, vector<int> input_shape, onnx::GraphProto *graph);
