+++ src/serialization/onnx/net/export_helpers.cpp
@@ -1,40 +1,7 @@
 #if defined(cPROTO)
 #include "eddl/serialization/onnx/export_helpers.h"
-
-#include "eddl/serialization/onnx/layers/core/dense_onnx.h"
-#include "eddl/serialization/onnx/layers/core/reshape_onnx.h"
-#include "eddl/serialization/onnx/layers/core/permute_onnx.h"
-#include "eddl/serialization/onnx/layers/core/activation_onnx.h"
-#include "eddl/serialization/onnx/layers/core/drop_onnx.h"
-#include "eddl/serialization/onnx/layers/core/squeeze_onnx.h"
-#include "eddl/serialization/onnx/layers/core/unsqueeze_onnx.h"
-#include "eddl/serialization/onnx/layers/core/embedding_onnx.h"
-#include "eddl/serialization/onnx/layers/merge/concat_onnx.h"
-#include "eddl/serialization/onnx/layers/merge/add_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/abs_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/diff_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/div_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/exp_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/log_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/mult_onnx.h"
-#include "eddl/serialization/onnx/layers/operators/sqrt_onnx.h"
-#include "eddl/serialization/onnx/layers/reductions/mean_onnx.h"
-#include "eddl/serialization/onnx/layers/reductions/sum_onnx.h"
-#include "eddl/serialization/onnx/layers/reductions/max_onnx.h"
-#include "eddl/serialization/onnx/layers/reductions/min_onnx.h"
-#include "eddl/serialization/onnx/layers/reductions/argmax_onnx.h"
-#include "eddl/serialization/onnx/layers/normalization/batchnorm_onnx.h"
-#include "eddl/serialization/onnx/layers/conv/conv_onnx.h"
-#include "eddl/serialization/onnx/layers/conv/conv1D_onnx.h"
-#include "eddl/serialization/onnx/layers/conv/upsampling_onnx.h"
-#include "eddl/serialization/onnx/layers/pool/maxpool_onnx.h"
-#include "eddl/serialization/onnx/layers/pool/maxpool1D_onnx.h"
-#include "eddl/serialization/onnx/layers/pool/avgpool_onnx.h"
-#include "eddl/serialization/onnx/layers/recurrent/lstm_onnx.h"
-#include "eddl/serialization/onnx/layers/recurrent/gru_onnx.h"
-#include "eddl/serialization/onnx/layers/recurrent/rnn_onnx.h"
-#include "eddl/serialization/onnx/layers/recurrent/cps_onnx.h"
-#include "eddl/serialization/onnx/layers/da/scale_onnx.h"
+#include "eddl/serialization/onnx/layers/layers_onnx.h"
+#include "eddl/layers/core/layer_core.h"
 
 onnx::ModelProto build_onnx_model(Net *net, bool gradients)
 {
@@ -175,119 +142,6 @@ void set_graph(onnx::ModelProto *model, Net *net, bool gradients)
   {
     // Builds a node of the graph from the layer in EDDL
     build_node_from_layer(aux_layer, graph, gradients, is_recurrent);
-  }
-}
-
-void build_node_from_layer(Layer *layer, onnx::GraphProto *graph, bool gradients, bool is_recurrent)
-{
-  // Check the class of the layer to call the corresponding function to
-  // build the node
-  if (LInput *l = dynamic_cast<LInput *>(layer))
-    return; // Skip the input layers
-  else if (LConv *l = dynamic_cast<LConv *>(layer))
-    build_conv_node(l, graph, gradients);
-  else if (LConv1D *l = dynamic_cast<LConv1D *>(layer))
-    build_conv1D_node(l, graph, gradients);
-  else if (LDense *l = dynamic_cast<LDense *>(layer))
-    if (is_recurrent)
-      build_dense_with_matmul_node(l, graph, gradients);
-    else
-      build_gemm_node(l, graph, gradients);
-  else if (LMaxPool *l = dynamic_cast<LMaxPool *>(layer))
-    build_maxpool_node(l, graph);
-  else if (LMaxPool1D *l = dynamic_cast<LMaxPool1D *>(layer))
-    build_maxpool1D_node(l, graph);
-  else if (LAveragePool *l = dynamic_cast<LAveragePool *>(layer))
-    build_averagepool_node(l, graph);
-  else if (LReshape *l = dynamic_cast<LReshape *>(layer))
-    build_reshape_node(l, graph);
-  else if (LPermute *l = dynamic_cast<LPermute *>(layer))
-    build_permute_node(l, graph);
-  else if (LUpSampling *l = dynamic_cast<LUpSampling *>(layer))
-    build_upsample_node(l, graph);
-  else if (LActivation *l = dynamic_cast<LActivation *>(layer))
-    // Check the type of activation layer
-    if (!l->act.compare("relu"))
-      build_relu_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("sigmoid"))
-      build_sigmoid_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("hard_sigmoid"))
-      build_hard_sigmoid_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("tanh"))
-      build_tanh_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("exp"))
-      build_exponential_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("linear"))
-      build_linear_node((LActivation *)(LinLayer *)layer, graph);
-    else if (!l->act.compare("leaky_relu"))
-      build_leaky_relu_node(l, graph);
-    else if (!l->act.compare("thresholded_relu"))
-      build_thresholded_relu_node(l, graph);
-    else if (!l->act.compare("elu"))
-      build_elu_node(l, graph);
-    else if (!l->act.compare("selu"))
-      build_selu_node(l, graph);
-    else if (!l->act.compare("softmax") || !l->act.compare("full_softmax"))
-      build_softmax_node(l, graph);
-    else if (!l->act.compare("softsign"))
-      build_softsign_node(l, graph);
-    else if (!l->act.compare("softplus"))
-      build_softplus_node(l, graph);
-    else 
-    {
-      cout << "The activation layer " << layer->name << "has no valid type to export." << endl;
-      return;
-    }
-  else if (LConcat *l = dynamic_cast<LConcat *>(layer))
-    build_concat_node(l, graph);
-  else if (LAbs *l = dynamic_cast<LAbs *>(layer))
-    build_abs_node(l, graph);
-  else if (LAdd *l = dynamic_cast<LAdd *>(layer))
-    build_add_node(l, graph);
-  else if (LDiv *l = dynamic_cast<LDiv *>(layer))
-    build_div_node(l, graph);
-  else if (LExp *l = dynamic_cast<LExp *>(layer))
-    build_exp_node(l, graph);
-  else if (LLog *l = dynamic_cast<LLog *>(layer))
-    build_log_node(l, graph);
-  else if (LMult *l = dynamic_cast<LMult *>(layer))
-    build_mul_node(l, graph);
-  //else if (LPow *l = dynamic_cast<LPow *>(layer))
-  //  build_pow_node(l, graph);
-  else if (LSqrt *l = dynamic_cast<LSqrt *>(layer))
-    build_sqrt_node(l, graph);
-  else if (LDiff *l = dynamic_cast<LDiff *>(layer))
-    build_sub_node(l, graph);
-  else if (LRMean *l = dynamic_cast<LRMean *>(layer))
-    build_rmean_node(l, graph);
-  else if (LRSum *l = dynamic_cast<LRSum *>(layer))
-    build_rsum_node(l, graph);
-  else if (LRMax *l = dynamic_cast<LRMax *>(layer))
-    build_rmax_node(l, graph);
-  else if (LRMin *l = dynamic_cast<LRMin *>(layer))
-    build_rmin_node(l, graph);
-  else if (LRArgmax *l = dynamic_cast<LRArgmax *>(layer))
-    build_rargmax_node(l, graph);
-  else if (LBatchNorm *l = dynamic_cast<LBatchNorm *>(layer))
-    build_batchnorm_node(l, graph);
-  else if (LDropout *l = dynamic_cast<LDropout *>(layer))
-    build_dropout_node(l, graph);
-  else if (LLSTM *l = dynamic_cast<LLSTM *>(layer))
-    build_lstm_node(l, graph);
-  else if (LGRU *l = dynamic_cast<LGRU *>(layer))
-    build_gru_node(l, graph);
-  else if (LRNN *l = dynamic_cast<LRNN *>(layer))
-    build_rnn_node(l, graph);
-  else if (LCopyStates *l = dynamic_cast<LCopyStates *>(layer))
-    handle_copy_states(l, graph);
-  else if (LEmbedding *l = dynamic_cast<LEmbedding *>(layer))
-    build_embedding_node(l, graph);
-  else if (LScale *l = dynamic_cast<LScale *>(layer))
-    build_resize_node(l, graph);
-  else
-  {
-    cout << "The layer " << layer->name << "has no OpType in Onnx." << endl;
-    return;
   }
 }
